+++ src/components/sections/elements/Services.vue
@@ -28,7 +28,7 @@ watch(
 <template>
   <SectionBorder>
     <SectionTitle
-      version="3.0"
+      version="4.0"
       :stability="Stability.Experimental"
       anchor="services"
       title="Services"