+++ src/lib/__tests__/distribution.spec.ts
@@ -1,67 +1,52 @@
 import { describe, it, expect } from 'vitest'
 
 import { ResourceType } from '@/types/enum'
-import type { Format, Licence } from '@/types/app'
+import type { Format, Licence, Organisation } from '@/types/app'
+import type { DistributionOption, PointOfContact as IsoContact } from '@/types/iso'
+import { getFormatByType, getOrganisation } from '@/lib/data'
+import { createOrgPointOfContact } from '@/lib/contacts'
 
 import {
   createDistributor,
-  getDistributorOrgSlug,
   getFileFormat,
   createDistributionOption,
   createDownloadDistributionOption,
   createServiceDistributionOption,
 } from '@/lib/distribution'
 
-const organisation = {
-  slug: 'bas',
-  name: 'British Antarctic Survey',
-  ror: 'https://ror.org/01rhff309',
-  phone: '+44 (0)1223 221400',
-  email: 'info@bas.ac.uk',
-  address: {
-    delivery_point: 'British Antarctic Survey, High Cross, Madingley Road',
-    city: 'Cambridge',
-    administrative_area: 'Cambridgeshire',
-    postal_code: 'CB3 0ET',
-    country: 'United Kingdom',
-  },
-  online_resource: {
-    href: 'https://www.bas.ac.uk',
-    title: 'British Antarctic Survey - BAS public website',
-    description: 'Homepage for the British Antarctic Survey (BAS) public website.',
-    function: 'information',
-  },
+const expectedOrganisation: Organisation = getOrganisation('bas_magic')
+
+const expectedFormat: Format = getFormatByType('image/png') as Format
+
+const expectedSizeBytes = 123
+
+const expectedDistributor: IsoContact = createOrgPointOfContact(expectedOrganisation, 'distributor')
+
+const expectedUrl = 'https://example.com'
+
+const expectedOnlineResource = {
+  href: expectedUrl,
+  title: 'PNG',
+  description: 'Download information as a PNG image',
+  function: 'download',
 }
 
-const expectedFormat: Format = {
-  slug: 'png',
-  name: 'PNG',
-  description: 'a PNG image',
-  extensions: ['.png'],
-  mediaTypes: ['image/png'],
-  url: 'https://www.iana.org/assignments/media-types/image/png',
-}
-
-const expectedDistributor = {
-  organisation: {
-    name: organisation.name,
-    href: organisation.ror,
-    title: 'ror',
+const expectedDistributionOption: DistributionOption = {
+  format: {
+    format: expectedFormat.name,
+    href: expectedFormat.url,
   },
-  phone: organisation.phone,
-  address: organisation.address,
-  email: organisation.email!,
-  online_resource: organisation.online_resource,
-  role: ['distributor'],
+  transfer_option: {
+    online_resource: expectedOnlineResource,
+    size: {
+      magnitude: expectedSizeBytes,
+      unit: 'bytes',
+    },
+  },
+  distributor: expectedDistributor,
 }
 
 describe('createDistributor', () => {
-  it('builds a distributor from an organisation', () => {
-    expect(createDistributor(organisation)).toStrictEqual(expectedDistributor)
-  })
-})
-
-describe('getDistributorOrgSlug', () => {
   const openLicence: Licence = {
     slug: 'x',
     name: 'x',
@@ -77,25 +62,25 @@ describe('getDistributorOrgSlug', () => {
     url: 'x',
     statement: 'x',
   }
-
   it('returns PDC for open datasets', () => {
-    expect(getDistributorOrgSlug(ResourceType.Dataset, openLicence)).toBe('nerc_eds_pdc')
-  })
-
-  it('returns null for closed datasets', () => {
-    expect(getDistributorOrgSlug(ResourceType.Dataset, closedLicence)).toBe(null)
+    const distributor = createOrgPointOfContact(getOrganisation('nerc_eds_pdc'), 'distributor')
+    expect(createDistributor(ResourceType.Dataset, openLicence)).toStrictEqual(distributor)
   })
 
   it('returns MAGIC for open products', () => {
-    expect(getDistributorOrgSlug(ResourceType.Product, openLicence)).toBe('bas_magic')
+    expect(createDistributor(ResourceType.Product, openLicence)).toStrictEqual(expectedDistributor)
   })
 
   it('returns MAGIC for closed products', () => {
-    expect(getDistributorOrgSlug(ResourceType.Product, closedLicence)).toBe('bas_magic')
+    expect(createDistributor(ResourceType.Product, closedLicence)).toStrictEqual(
+      expectedDistributor
+    )
   })
 
   it('returns null for other resource types', () => {
-    expect(getDistributorOrgSlug(ResourceType.Collection, openLicence)).toBe(null)
+    expect(createDistributor(ResourceType.Collection, openLicence)).toStrictEqual(
+      expectedDistributor
+    )
   })
 })
 
@@ -150,125 +135,37 @@ describe('getFileFormat', () => {
 })
 
 describe('createDistributionOption', () => {
-  it('builds a distribution option from a format (without format), online resource and organisation', () => {
-    const format = expectedFormat
-    const onlineResource = {
-      href: 'https://example.com',
-      title: 'Example',
-      description: 'Example description',
-      function: 'download',
-    }
-    const expectedDistributionOption = {
-      format: {
-        format: format.name,
-        href: format.url,
-      },
-      transfer_option: {
-        online_resource: onlineResource,
-      },
-      distributor: expectedDistributor,
-    }
-
-    expect(createDistributionOption(format, onlineResource, organisation)).toStrictEqual(
-      expectedDistributionOption
-    )
+  it('builds a distribution option from a format, size, online resource and organisation', () => {
+    expect(
+      createDistributionOption(
+        expectedFormat,
+        expectedOnlineResource,
+        expectedDistributor,
+        expectedSizeBytes
+      )
+    ).toStrictEqual(expectedDistributionOption)
   })
 
-  it('builds a distribution option from a format (with format), online resource and organisation', () => {
-    const format = {
-      slug: 'gpkg',
-      name: 'GeoPackage',
-      version: '1.2',
-      extensions: ['.gpkg'],
-      mediaTypes: ['application/geopackage+sqlite3'],
-      url: 'https://www.iana.org/assignments/media-types/application/geopackage+sqlite3',
-    }
-    const onlineResource = {
-      href: 'https://example.com',
-      title: 'Example',
-      description: 'Example description',
-      function: 'download',
-    }
-    const expectedDistributionOption = {
-      format: {
-        format: format.name,
-        href: format.url,
-        version: format.version,
-      },
-      transfer_option: {
-        online_resource: onlineResource,
-      },
-      distributor: expectedDistributor,
-    }
+  it('builds a distribution option from a format, no size online resource, organisation and size', () => {
+    const distOptNoSize: DistributionOption = JSON.parse(JSON.stringify(expectedDistributionOption))
+    delete distOptNoSize.transfer_option.size
 
-    expect(createDistributionOption(format, onlineResource, organisation)).toStrictEqual(
-      expectedDistributionOption
-    )
-  })
-
-  it('builds a distribution option from a format, online resource, organisation and size', () => {
-    const format = expectedFormat
-    const onlineResource = {
-      href: 'https://example.com',
-      title: 'Example',
-      description: 'Example description',
-      function: 'download',
-    }
-    const sizeBytes = 123
-    const expectedDistributionOption = {
-      format: {
-        format: format.name,
-        href: format.url,
-      },
-      transfer_option: {
-        online_resource: onlineResource,
-        size: {
-          magnitude: sizeBytes,
-          unit: 'bytes',
-        },
-      },
-      distributor: expectedDistributor,
-    }
-
-    expect(createDistributionOption(format, onlineResource, organisation, sizeBytes)).toStrictEqual(
-      expectedDistributionOption
-    )
+    expect(
+      createDistributionOption(expectedFormat, expectedOnlineResource, expectedDistributor)
+    ).toStrictEqual(distOptNoSize)
   })
 })
 
 describe('createDownloadDistributionOption', () => {
-  it('builds a distribution option from a file, url and organisation', () => {
-    const file: File = {
-      name: 'foo.png',
-      type: 'image/png',
-      lastModified: 0,
-      webkitRelativePath: '',
-      size: 0,
-      arrayBuffer: () => Promise.resolve(new ArrayBuffer(0)),
-      slice: () => new Blob(),
-      stream: () => new ReadableStream(),
-      text: () => Promise.resolve(''),
-    }
-    const url = 'https://example.com'
-    const expectedDistributionOption = {
-      format: {
-        format: 'PNG',
-        href: 'https://www.iana.org/assignments/media-types/image/png',
-      },
-      transfer_option: {
-        online_resource: {
-          href: url,
-          title: 'PNG',
-          description: 'Download information as a PNG image',
-          function: 'download',
-        },
-      },
-      distributor: expectedDistributor,
-    }
-
-    expect(createDownloadDistributionOption(file, url, organisation)).toStrictEqual(
-      expectedDistributionOption
-    )
+  it('builds a distribution option from a format, url and organisation', () => {
+    expect(
+      createDownloadDistributionOption(
+        expectedFormat,
+        expectedUrl,
+        expectedDistributor,
+        expectedSizeBytes
+      )
+    ).toStrictEqual(expectedDistributionOption)
   })
 })
 
@@ -297,7 +194,7 @@ describe('createServiceDistributionOption', () => {
       distributor: expectedDistributor,
     }
 
-    expect(createServiceDistributionOption(service, url, organisation)).toStrictEqual(
+    expect(createServiceDistributionOption(service, url, expectedDistributor)).toStrictEqual(
       expectedDistributionOption
     )
   })